chore(refactor): move utils into root tools folder
