+++ gen3/tools/bundle/ingest_manifest.py
@@ -10,7 +10,7 @@ import re
 from drsclient.client import DrsClient
 from gen3.auth import Gen3Auth
 from gen3.utils import UUID_FORMAT, SIZE_FORMAT, _verify_format, _standardize_str
-from gen3.tools.indexing.manifest_columns import (
+from gen3.tools.utils import (
     GUID_COLUMN_NAMES,
     SIZE_COLUMN_NAMES,
     BUNDLENAME_COLUMN_NAME,
