+++ gen3/tools/indexing/download_manifest.py
@@ -33,8 +33,9 @@ import sys
 import shutil
 import math
 
+from gen3.tools.utils import get_and_verify_fileinfos_from_manifest, MD5_STANDARD_KEY
+from gen3.utils import get_or_create_event_loop_for_thread, yield_chunks
 from gen3.index import Gen3Index
-from gen3.utils import get_or_create_event_loop_for_thread
 
 INDEXD_RECORD_PAGE_SIZE = 1024
 MAX_CONCURRENT_REQUESTS = 24
