+++ gen3/tools/indexing/merge_manifests.py
@@ -30,8 +30,8 @@ import csv
 import copy
 
 from collections import OrderedDict
-from gen3.tools.indexing.index_manifest import get_and_verify_fileinfos_from_manifest
-from gen3.tools.indexing.manifest_columns import (
+from gen3.tools.utils import get_and_verify_fileinfos_from_manifest
+from gen3.tools.utils import (
     GUID_STANDARD_KEY,
     SIZE_STANDARD_KEY,
     MD5_STANDARD_KEY,
