+++ pyproject.toml
@@ -25,7 +25,6 @@ drsclient = "<1.0.0"
 aiohttp = "*"
 backoff = "*"
 click = "*"
-pandas = "*"
 jsonschema = "*"
 dataclasses-json = "*"
 pypfb = "<1.0.0"
